+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-11 14:44:28
-// Commit: fix(core): prevent double tap duplicate action in core
+// Updated at: 2026-08-11 17:04:49
+// Commit: fix(core): handle null assertion failure in core parser
 
 class coreComponentState {
-  final String commitRef = '3527';
-  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
-  final DateTime timestamp = DateTime.parse('2026-08-11T14:44:28+05:00');
+  final String commitRef = '3536';
+  final String lastMessage = 'fix(core): handle null assertion failure in core parser';
+  final DateTime timestamp = DateTime.parse('2026-08-11T17:04:49+05:00');
 }